requirements.txt
asgiref==3.11.0
Django==6.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.21
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3
